pyproject.toml
[project]
name = "dataset-parallel-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    #    "aiohttp",
    "datasets",
    "openai",
    "python-dotenv",
    "tqdm",
]
